+++ plugins/kernel/backend/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @webstackbuilders/plugin-ai-core-backend
+# @ai-crew-suite/plugin-kernel-backend
 
 ## 0.0.1